pyproject.toml
[project]
name = "legal-spend-mcp"
version = "0.1.0"
description = "MCP server for legal spend intelligence across multiple data sources"
authors = [
    {name = "DatSciX", email = "patrick@datscix.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
    "pymssql>=2.2.0",
    "cx_Oracle>=8.3.0",
    "psycopg2-binary>=2.9.0",
    "openpyxl>=3.1.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.scripts]
legal-spend-mcp = "legal_spend_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310"]